improved object removal scheme, fix mem leaks
memory is now freed properly, and removing objects is a more flehsed out process. Also there is a difference between ALIVE and EXISTS now. Alive is just, "is it queued to be destroyed", exists means "is there data here, or can I write to this buffer"
You can now queue component lists based on the alive flag, this is so that ccomponent heap memory can be freed before the pointer to it is lost. This will take place in the post logic systems